\name{integrativeME-internal}
\encoding{latin1}
\alias{norm.indep}
\alias{initmain.indep}
\alias{main.indep}
\alias{norm.logreg}
\alias{initmain.logreg}
\alias{main.logreg}
\alias{norm.loc}
\alias{initmain.loc}
\alias{main.loc}
\alias{counts}

\title{
internal integrativeMe functions
}
\description{
Internal functions in integrativeME}


\details{ These functions are not to be called by the user.
}

\author{Kim-Anh Le Cao}



