\name{kmeansME}
\encoding{latin1}
\alias{kmeansME}
\title{
independence gating function for integrative ME
}
\description{
More about what it does (maybe more than one line)
}


%\usage{}

%\arguments{
%  \item{}{}
%}

%\details{}

%\value{}

%\references{}

%\author{}

%\seealso{}

%\examples{}


